Use CGFloat

+++ Classes/WOS7ListApp.m
@@ -1,6 +1,6 @@
 #import "WOS7ListApp.h"
 
-#define UIColorFromRGB(rgbValue) [UIColor colorWithRed:((float)((rgbValue & 0xFF0000) >> 16))/255.0 green:((float)((rgbValue & 0xFF00) >> 8))/255.0 blue:((float)(rgbValue & 0xFF))/255.0 alpha:1.0]
+#define UIColorFromRGB(rgbValue) [UIColor colorWithRed:((CGFloat)((rgbValue & 0xFF0000) >> 16))/255 green:((CGFloat)((rgbValue & 0xFF00) >> 8))/255 blue:((CGFloat)(rgbValue & 0xFF))/255 alpha:1]
 
 @implementation WOS7ListApp
